Fix scan mode switch issue
Resolved an issue where switching to scan mode failed due to incorrect
interface handling.

If the platform uses wlan0 for both STA and NAN (or type other type),
and the existing iface list in WifiNative IfaceManager is

{(wlan0, NAN), (wlan0 STA_CONNECT)}

in order, when WifiNative tries to switch STA to scan only mode, it will
mistakenly change (wlan0, NAN) to (wlan0, STA_SCAN) and produces a list
containing

{(wlan0, STA_SCAN), (wlan0, STA_CONNECTIVITY)}

and causes chaos later. User will not be able to enable Wi-Fi.

Flag: EXEMPT bugfix
Bug: 395498484
Test: Manually test STA enable/disable.
      Automatic tests of Aware publish/subscribe.

Change-Id: I76138f8c7e9de711a2f3ac7d5d3e506da4cab7f8