Move to GroupCreatedState after user accepting the connect request
After receiving the PEER_CONNECTION_USER_ACCEPT event, the state machine UserAuthorizingJoinState
did not switch to GroupCreatedState.
As a result, when the third device was connected, it could not switch from GroupCreatedState
to UserAuthorizingJoinState, and the Dialog could not be popped up to click Accept.

Bug: 397272526
Test: Manual - P2P connection via WFD window
Test: Presubmit tests
Flag: com.android.wifi.flags.wifi_direct_r2
Change-Id: I4ea6333367fe03d6b7b7858d13bdce3598fdc2be