Notify WifiStateChangedListener upon registration
Clients of WifiStateChangedListener must use WifiManager.getWifiState()
to query the current wifi state, but this state might change between
WifiManager.addWifiStateChangedListener and the listener actually
registering on the Wifi thread, which the client will not be notified
for.

Fix this by calling WifiStateChangedListener.onWifiStateChanged
immediately after the listener is registered on the wifi thread. This
will allow apps to trust that any changes to WifiManager.getWifiState()
afterwards will notify the listener.

Flag: com.android.wifi.flags.wifi_state_changed_listener

Bug: 393488380
Test: atest ActiveModeWardenTest
Change-Id: Iaddae488c3e3c0ee67949bfd4d9ec5d15af67be5