WiFi: Force 11be to false when 11ax is false.
The 11be hotspot requires 11ax due to a dependency.
When 11ax is false, but 11be is configured, it can cause
a hostapd error (e.g., b/378048737). To prevent this,
especially when OEMs might retain invalid SoftApConfigurations
in Settings, we should implement the following protection:

Protection for Android 16 and later.
Automatically force 11be to false when 11ax is false.
Trigger an error for invalid configurations,
preventing changes in the Settings
(this applies only when callers bypass the SDK API,
such as through reflection).

Flag: EXEMPT minor fix
Bug: 401184848
Test: atest -c FrameworksWifiTests
Test: atest -c FrameworksWifiApiTests
Change-Id: I5b4ebc2a3dd82c02bb0f742815e3b13f40444c4c