Update RSSI when supplicant connects
Verbose summary in the Wifi picker will show the WifiInfo rssi upon
association, but this may be -127 for a short time until RSSI polling
starts. Make this more responsive by updating the WifiInfo rssi when
supplicant associates.

Note that this only affects the RSSI displayed in the verbose summary,
and not the wifi pie, which will still be showing the rssi from scan
results.

Flag: EXEMPT debug log change

Bug: 399407645
Test: atest ClientModeImplTest
Change-Id: Ieffacc00e3a28fc90e73c6297201d4be43e2a8b9