Refactor SimpleDialogs to use a builder
SimpleDialogs currently use a constructor method with several arguments,
which makes it difficult to add new arguments in the future. Instead,
add a SimpleDialogBuilder to simplify the construction.

This CL only migrates WifiShellCommand to use the builder. The rest of
the clients will start using the builder in a future CL.

Flag: EXEMPT mechanical refactor

Bug: 276778134
Test: build, manually verify with adb shell cmd wifi
launch-dialog-simple.

Change-Id: I012db003edfd24c2a2fabfe6d0f80170ff16fb62