Allow non-enterprise WiFi suggestions on debug builds
Relax the check in WifiNetworkSuggestion#Builder to allow
non-user builds to add carrier-merged suggestions that ar
non-enterprise so long as they comply with the other rules
(like being metered and attached to a subscription.)
This allows simpler internal testing of carrier merged Wifi
networks, but will not be deployable for release builds.

Note that this is not totally ideal, but since all
authoritative range checking is done on the server side,
this does not create a functional regression, it just
allows someone to build a suggestion that may later be
rejected if it violates the policy.

Bug: 408451862
Test: atest WifiNetworkSuggestionTest
Flag: EXEMPT bugfix
Change-Id: Id223c0a3a0636f7f4796505ddbb66e19c0f7273c