Fix WifiP2pConfig Parcelable backward compatibility.
The `mPccModeConnectionType` field, added in Android 16, was inserted
into the middle of the `Parcelable` sequence in `writeToParcel`
and `createFromParcel`. This caused data corruption when older applications
unparceled `WifiP2pConfig` objects created on newer versions.

Move the read/write operations for `mPccModeConnectionType` to the end
of the sequence to restore compatibility.

Bug: 413549603
Flag: com.android.wifi.flags.wifi_direct_r2
Test: Manual - P2P connection
Test: TH Presubmit tests
Test: atest com.android.server.wifi.p2p
Change-Id: I10e37813f8ded5ece022c99578f797abd4fa1da0