Wifi: WiFi Configuration: Enhancing User Tracking
To improve multi-user support, a new API is requested for WiFi configurations.
Currently, the "creatorUid" (application UID) does not reveal the creator's
user ID when the UID belongs to the system. This proposal suggests adding a
"creator user ID" field to the logging process when a new WiFi configuration
is created. This will enable proper identification of the actual user
responsible for the configuration, addressing the existing limitation.

Flag: com.android.wifi.flags.multi_user_wifi_enhancement
Bug: 390240471
Test: attest -c FrameworksWifiApiTests / FrameworksWifiTests
Reviewed-Cloud-B&R: TRUE

Change-Id: I3bd1bec03f8010e28f11e5bf40be72183577069b