API for WFDR2 external approver support
This CL include the changes for,
1. Convert getPairingBootstrappingMethod() hide API to public
for external approver on the responder device to get the
bootstrapping method
2. Convert getPairingBootstrappingPassword() hide API to public
for external approver on the responder device to get the
password from WifiP2pConfig for showing it on UI or to send out
of band to the other device.
3. Added a new callback called onPasswordGenerated
to notify external approver that a password is generated.
4. Modified the documentation of WifiP2pManager#setConnectionRequestResult()
documentation to indicate that the API can be used for
setting the password for WFD R2 connection.
5. Added a new constant CONNECTION_REQUEST_DEFER_SHOW_PASSWORD_TO_SERVICE
to let the service implmentation to display the password

Bug: 409802099
Flag: com.android.wifi.flags.external_approver_support_for_wfdr2_password_based_bootstrapping
Test: Manual - Basic P2P connect/disconnect tests
Change-Id: Icfa3077d4f92a083656e62b74da6dbedc0131f7b