Centralize logic for roaming control
When we want to force the station to only connect to a particular BSSID,
we need to call a HAL command to disable roaming.

Today, we are calling this command at multiple different places before
we make the connection. This is error-prone, and there's a bug that we
are missing disabling roaming when manually forcing connection to a
saved network. This is causing flaky CTS failure.

Centralize the logic to disable roaming based on forcing BSSID so that
we don't miss anything in the future.

Flag: EXEMPT cts test fix
Bug: 421040163
Test: atest WifiManagerTest
Change-Id: If9280df91684884f1fc77e67e5b6bf21558ba392