Fix MIN_RTT_BURST_SIZE to 1
RTT burst size is applicable to IEEE 802.11mc, and for one special case
it is also applicable to IEEE 802.11az to generate multiple NTB ranging
requests per measurement. It is applicable for IEEE 802.11az based
ranging requests when MIMO is not available, with the transmit and
receive spatial streams between the initiator and responder station is
equal to 1.

For IEEE 802.11mc the minimum burst size is 2 and for IEEE 802.11az NTB
measurement the minimum burst size is 1.

Since a ranging request can have a mix of 11mc and 11az peers, make the
minimum value for RTT burst size to 1.

Bug: 414866300
Flag: EXEMPT bug fix
Test: atest FrameworksWifiApiTests
Change-Id: I78aeade59b396859f6c677e31358b3602c843dbd