Fix NullPointerException in background scan getScanResults
It is possible that the `getScanResults()` API
dispatches a `CMD_GET_SCAN_RESULTS` message to the
`WifiBackgroundScanStateMachine` without an object payload
(`msg.obj` is null).

If the state machine is in its `DefaultState`, the message handler
attempts to cast the null `msg.obj` to
`ScanParams` and then immediately dereferences it to access
`scanParams.listener`. This results in a `NullPointerException`.

Fix this by adding a null check for the `ScanParams` object within the
`DefaultState`'s message handler.

Bug: 422706327
Flag: EXEMPT bug fix
Test: Presubmit tests
Change-Id: I2cd504d3013323c764c5e3aa37ae304a4d8f4f31