Avoid calling getCachedScanData if unsupported
Avoid calling the `getCachedScanData()` HAL API if the
device reports that it is not supported.

This is fixed by assuming the feature is supported on startup.
If a call to the HAL returns a `ServiceSpecificException` with
the error code `ERROR_NOT_SUPPORTED`, the feature is marked as
unsupported to prevent any subsequent calls.

Bug: 428876881
Flag: EXEMPT bug fix
Test: atest WifiStaIfaceAidlImplTest
Change-Id: I86e45f7d042e74ca9aa23e142e2492b53a10f095