wifip2p: catch exception for unregisterTetheringEventCallback
A call to TetheringManager#unregisterTetheringEventCallback
may throw an exception if the callback somehow removed from
tethering module or if the tethering service is not available during teardown.

It is safe to ignore such exception. So catch the exception and continue
processing the P2P state change.

Bug: 432355563
Flag: EXEMPT bug fix
Test: atest WifiP2pServiceImplTest
Test: Manul - Wi-Fi Direct ON/OFF
Change-Id: Ib6b77ac4e83b446d9de275d59996e5447214525d