Require user approval for WFD-R2 Display PIN connections
When receiving a Wi-Fi Direct R2 (WFD-R2) connection request using the
'Display PIN/Password' bootstrapping method on the group owner,
the framework would immediately authorize the connection in the supplicant.

This bypassed user consent, as the connection was authorized before the
user could interact with the confirmation dialog.

This change fixes the issue by transitioning to the `UserAuthorizingJoinState`
upon receiving the request. In this state, a dialog is presented to the
user. The connection is only authorized after the user explicitly accepts,
ensuring no connection is established without consent.

Bug: 439658889
Flag: com.android.wifi.flags.wifi_direct_r2
Test: atest WifiP2pServiceImplTest
Change-Id: Id439f9519dd5d8efe5895c9d19c449ba0c4b7d10