Drop the Transition Disable event if WPA3-SAE is not supported by the device.
According to the specification, an Access Point (AP) should not send a
Transition Disable indication for WPA2 or WPA associations.
However, some APs fail to adhere to this, sending the indication regardless.
On legacy devices that don't support WPA3-SAE, this behavior causes a bug
where the framework mistakenly disables WPA2, thereby preventing future
connections to that network.

This change resolves the problem by ensuring the framework only processes
the SAE Transition Disable event if the device has support for SAE and SAE upgrade.

Additionaly this change also fixes the potential issue in OWE and WPA3-Enterprise
by checking the security parameters of the last successful connection
(`getLastUsedSecurityParams()`). The framework now only disables a weaker security
type if it can verify that the stronger security type was actually used for the most
recent connection.

Bug: 441400904
Flag: EXEMPT bug fix
Test: atest ClientModeImplTest
Test: atest Manual - Connect to an AP operating in WPA3-TM with TDI
      enabled and confirmed from the logs that framework is handling the TDI
      as per the new logic.
Change-Id: Id7cb4d3b02055f8fe88f3acdbb1966c867af695b