Enable Frame Protection for Aware data paths when pairing is used.
This change introduces a mechanism to enable Frame Protection for Wi-Fi Aware data paths. The `WifiAwareDiscoverySessionState` now tracks peers that have successfully completed the pairing confirmation. When initiating or responding to a data path request, the system checks if the peer is paired. If the peer is paired, a `frameProtectionEnabled` flag is set and passed down to the HAL layer. The HAL implementation then configures the data path security to require frame protection, including support for BIGTKSA, GTKSA, and IGTKSA.

Bug: 440147300
Test: atest com.android.server.wifi
Flag: EXEMPT BUGFIX
Change-Id: I554ff19c96ae6ab1c1ce7a088fbc4adbbde03ae6